+++ src/main/java/duke/data/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.data;
 
 import java.util.ArrayList;
 import duke.task.Task;
