+++ src/main/java/duke/parser/Parser.java
@@ -1,15 +1,17 @@
-package duke;
+package duke.parser;
 
+import duke.storage.Storage;
+import duke.data.TaskList;
 import duke.exception.*;
 import duke.task.Task;
 import duke.task.ToDo;
 import duke.task.Event;
 import duke.task.Deadline;
+import duke.ui.Ui;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-import java.util.Date;
 
 /**
  * Deals with making sense of user command
@@ -40,15 +42,21 @@ public class Parser {
     public String parse(String input) throws DukeException {
                if (input.equals("bye")) {
                    return ui.bye();
-               } else if (input.equals("list")) {
+               }
+
+               if (input.equals("list")) {
                    return ui.showTaskList(this.tasks, "list");
-               } else if (input.contains("done")) {
+               }
+
+               if (input.contains("done")) {
                    int taskNum = getNum(input);
                    Task task = tasks.getTask(taskNum);
                    task.markAsDone();
                    storage.updateData(tasks);
                    return ui.showTaskDone(task);
-               } else if (input.contains("todo")) {
+               }
+
+               if (input.contains("todo")) {
                    String name = getName(input);
                    if (!name.equals("")) {
                        ToDo toDo = new ToDo(name);
@@ -56,10 +64,12 @@ public class Parser {
                        storage.updateData(tasks);
                        return ui.showTaskAdded(toDo, tasks.getSize());
                    }
-               } else if (input.contains("deadline")) {
+               }
+
+               if (input.contains("deadline")) {
                    String description = getName(input);
                    if (!description.equals("")) {
-                       if (descriptionIsValid(description, "/by", "deadline")) {
+                       if (checkDescriptionIsValid(description, "/by", "deadline")) {
                            String[] parts = description.split("/by");
                            String name = parts[0];
                            LocalDateTime by = getDateTime(parts[1]);
@@ -70,10 +80,12 @@ public class Parser {
                            return ui.showTaskAdded(deadline, tasks.getSize());
                        }
                    }
-               } else if (input.contains("event")) {
+               }
+
+               if (input.contains("event")) {
                    String description = getName(input);
                    if (!description.equals("")) {
-                       if (descriptionIsValid(description, "/at", "event")) {
+                       if (checkDescriptionIsValid(description, "/at", "event")) {
                            String[] parts = description.split("/at");
                            String name = parts[0];
                            LocalDateTime at = getDateTime(parts[1]);
@@ -84,13 +96,17 @@ public class Parser {
                            return ui.showTaskAdded(event, tasks.getSize());
                        }
                    }
-               } else if (input.contains("delete")) {
+               }
+
+               if (input.contains("delete")) {
                    int taskNum = getNum(input);
                    Task task = tasks.getTask(taskNum);
                    tasks.deleteTask(task);
                    storage.updateData(tasks);
                    return ui.showTaskDeleted(task, tasks.getSize());
-               } else if (input.contains("find")) {
+               }
+
+               if (input.contains("find")) {
                    String keyword = getName(input);
                    if (!keyword.equals("")) {
                        TaskList t = tasks.matchTasks(keyword);
@@ -149,23 +165,25 @@ public class Parser {
         }
     }
 
-    private boolean descriptionIsValid(String description, String pre, String task)
+    private boolean checkDescriptionIsValid(String description, String pre, String task)
             throws TaskNoDateTimeException, TaskNoNameException, EmptyDescriptionException, MissingPreException {
         if (description.contains(pre)) {
-            if (!description.equals(pre)
-                    && !(description.startsWith(pre) && description.endsWith(" "))
-                    && !(description.startsWith(" ") && description.endsWith(pre))) {
+            if (checkForMissingDescription(description, pre)) {
                 if(description.startsWith(pre)) {
                     throw new TaskNoNameException("error", task);
                 } else if (description.endsWith(pre)) {
                     throw new TaskNoDateTimeException("error", task);
                 }
                 return true;
-            } else {
-                throw new EmptyDescriptionException("error", task);
             }
-        } else {
-            throw new MissingPreException("error", pre);
+            throw new EmptyDescriptionException("error", task);
         }
+        throw new MissingPreException("error", pre);
+    }
+
+    boolean checkForMissingDescription(String description, String pre) {
+        return !description.equals(pre)
+                && !(description.startsWith(pre) && description.endsWith(" "))
+                && !(description.startsWith(" ") && description.endsWith(pre));
     }
 }
